feature: distance, duration 단위 맞추기/마커 clickListener 추가

+++ app/src/main/java/com/example/picknumberproject/dto/directions5/Summary.kt
@@ -3,7 +3,7 @@ package com.example.picknumberproject.dto.directions5
 data class Summary(
     val bbox: List<List<Double>>,
     val departureTime: String,
-    val distance: Int,
+    val distance: Double,
     val duration: Int,
     val etaServiceType: Int,
     val fuelPrice: Int,
